Replaced floating point sqrt() with integer sqrt(). This is temporary needed because the FPU is not properly initialized/context switched for the application processors on Intel. Modified MemoryChannel::write() to busy wait in case of a full channel. Renamed prime to mpiprime program for parallel prime number computation. Added prime program for sequential prime number computation.

+++ lib/libposix/math.h
@@ -43,7 +43,7 @@ extern C int abs(int number);
  * @param number The number to use for square root.
  * @return Square root of the given number.
  */
-extern C double sqrt(double number);
+extern C u32 sqrt(u32 number);
 
 /**
  * @}
